remove stage package
